+++ Natives/JavaGUIViewController.m
@@ -3,12 +3,12 @@
 #import "JavaGUIViewController.h"
 #import "JavaLauncher.h"
 #import "LauncherPreferences.h"
-#import "MousePointerFactory.h"
 #import "PLLogOutputView.h"
 #import "TrackedTextField.h"
 #import "UnzipKit.h"
 #import "authenticator/BaseAuthenticator.h"
 #import "ios_uikit_bridge.h"
+#import "CursorManager.h"
 #include "glfw_keycodes.h"
 #include "utils.h"
 
@@ -70,10 +70,10 @@ void AWTInputBridge_sendKey(int keycode) {
         return;
     } else if (!surfaceJNIEnv) {
         // Obtain JNIEnvs
-        (*runtimeJavaVMPtr)->AttachCurrentThread(runtimeJavaVMPtr, &surfaceJNIEnv, NULL);
+        (*runtimeJavaVMPtr)->AttachCurrentThreadAsDaemon(runtimeJavaVMPtr, &surfaceJNIEnv, NULL);
         if (!surfaceJNIEnv) return;
         dispatch_async(dispatch_get_main_queue(), ^{
-            (*runtimeJavaVMPtr)->AttachCurrentThread(runtimeJavaVMPtr, &runtimeJNIEnvPtr, NULL);
+            (*runtimeJavaVMPtr)->AttachCurrentThreadAsDaemon(runtimeJavaVMPtr, &runtimeJNIEnvPtr, NULL);
             if (!runtimeJNIEnvPtr) return;
         });
 
@@ -134,7 +134,6 @@ void AWTInputBridge_sendKey(int keycode) {
 
 @interface ScrollableSurfaceView<UIScrollViewDelegate> : UIScrollView
 @property CGRect clickRange, virtualMouseFrame;
-@property CGPoint mouseHotspot;
 @property(nonatomic) UIImageView* mousePointerView;
 @property BOOL shouldTriggerClick;
 @end
@@ -150,12 +149,9 @@ void AWTInputBridge_sendKey(int keycode) {
     self.virtualMouseFrame = CGRectMake(frame.size.width / 2, frame.size.height / 2, 18, 27);
     self.mousePointerView = [[UIImageView alloc] initWithFrame:self.virtualMouseFrame];
     self.mousePointerView.hidden = !virtualMouseEnabled;
-    self.mousePointerView.contentMode = UIViewContentModeScaleAspectFit;
-    id pointerStyleObj = getPrefObject(@"control.mouse_pointer_style");
-    NSString *pointerStyle = [pointerStyleObj isKindOfClass:[NSString class]] ? pointerStyleObj : @"default";
-    self.mousePointerView.image = [MousePointerFactory imageForStyle:pointerStyle];
-    self.mouseHotspot = [MousePointerFactory hotspotForStyle:pointerStyle];
+    self.mousePointerView.image = [CursorManager imageForCursor:[CursorManager currentCursorName]];
     [surfaceView addSubview:self.mousePointerView];
+    self.mousePointerView.frame = [CursorManager displayFrameForMouseFrame:self.virtualMouseFrame];
 
     return self;
 }
@@ -180,13 +176,11 @@ void AWTInputBridge_sendKey(int keycode) {
         // Calculate delta
         location.x = (location.x - prevLocation.x) / self.zoomScale;
         location.y = (location.y - prevLocation.y) / self.zoomScale;
-        // Update cursor's position
-        CGFloat cursorX = clamp(self.virtualMouseFrame.origin.x + location.x, 0, self.frame.size.width * self.zoomScale);
-        CGFloat cursorY = clamp(self.virtualMouseFrame.origin.y + location.y, 0, self.frame.size.height * self.zoomScale);
-        _virtualMouseFrame.origin.x = cursorX - self.virtualMouseFrame.size.width * self.mouseHotspot.x;
-        _virtualMouseFrame.origin.y = cursorY - self.virtualMouseFrame.size.height * self.mouseHotspot.y;
-        self.mousePointerView.frame = self.virtualMouseFrame;
-        location = CGPointMake(cursorX, cursorY);
+        // Update cursor's origin
+        _virtualMouseFrame.origin.x = clamp(self.virtualMouseFrame.origin.x + location.x, 0, self.frame.size.width * self.zoomScale);
+        _virtualMouseFrame.origin.y = clamp(self.virtualMouseFrame.origin.y + location.y, 0, self.frame.size.height * self.zoomScale);
+        self.mousePointerView.frame = [CursorManager displayFrameForMouseFrame:self.virtualMouseFrame];
+        location = self.virtualMouseFrame.origin;
 
         CGPoint minimumContentOffset = CGPointMake(-self.contentInset.left, -self.contentInset.top);
         CGPoint maximumContentOffset = CGPointMake(
@@ -194,8 +188,8 @@ void AWTInputBridge_sendKey(int keycode) {
             MAX(minimumContentOffset.y, self.contentSize.height + self.contentInset.bottom - self.frame.size.height));
         // Focus scroll view's content area on virtual mouse
         self.contentOffset = CGPointMake(
-            clamp(location.x * self.zoomScale - self.center.x, minimumContentOffset.x, maximumContentOffset.x),
-            clamp(location.y * self.zoomScale - self.center.y, minimumContentOffset.y, maximumContentOffset.y));
+            clamp(self.virtualMouseFrame.origin.x * self.zoomScale - self.center.x, minimumContentOffset.x, maximumContentOffset.x),
+            clamp(self.virtualMouseFrame.origin.y * self.zoomScale - self.center.y, minimumContentOffset.y, maximumContentOffset.y));
     }
 
     // Send cursor position to AWT
@@ -206,14 +200,12 @@ void AWTInputBridge_sendKey(int keycode) {
 - (void)scrollViewDidZoom:(UIScrollView *)scrollView {
     if (virtualMouseEnabled) {
         // Keep virtual mouse in the middle of screen while zooming
-        CGFloat cursorX = (self.contentOffset.x + self.center.x) / self.zoomScale;
-        CGFloat cursorY = (self.contentOffset.y + self.center.y) / self.zoomScale;
-        _virtualMouseFrame.origin.x = cursorX - self.virtualMouseFrame.size.width * self.mouseHotspot.x;
-        _virtualMouseFrame.origin.y = cursorY - self.virtualMouseFrame.size.height * self.mouseHotspot.y;
-        self.mousePointerView.frame = self.virtualMouseFrame;
+        _virtualMouseFrame.origin.x = (self.contentOffset.x + self.center.x) / self.zoomScale;
+        _virtualMouseFrame.origin.y = (self.contentOffset.y + self.center.y) / self.zoomScale;
+        self.mousePointerView.frame = [CursorManager displayFrameForMouseFrame:self.virtualMouseFrame];
         // Send cursor position to AWT
         CGFloat screenScale = UIScreen.mainScreen.scale * getPrefFloat(@"video.resolution") / 100.0;
-        AWTInputBridge_nativeSendData(EVENT_TYPE_CURSOR_POS, (int)(cursorX * screenScale), (int)(cursorY * screenScale), 0, 0);
+        AWTInputBridge_nativeSendData(EVENT_TYPE_CURSOR_POS, (int)(_virtualMouseFrame.origin.x * screenScale), (int)(_virtualMouseFrame.origin.y * screenScale), 0, 0);
     }
 }