+++ Natives/egl_bridge.m
@@ -49,7 +49,6 @@ int pojavInit(BOOL useStackQueue) {
     clientAPI = GLFW_OPENGL_API;
     isInputReady = 1;
     isUseStackQueueCall = useStackQueue;
-    NSLog(@"[InputDiag] pojavInit: useStackQueue=%d (isInputReady set to 1)", useStackQueue);
     aasdl_setMainReady();
     return JNI_TRUE;
 }